moved pact-related functionality to PactProject
