+++ src/componentTest/groovy/com/blackbaud/templates/tasks/TemplateGenerationSpec.groovy
@@ -6,6 +6,7 @@ import com.blackbaud.templates.project.BasicProject
 import com.blackbaud.templates.project.EventHubsProject
 import com.blackbaud.templates.project.IntegrationTestProject
 import com.blackbaud.templates.project.KafkaProject
+import com.blackbaud.templates.project.PactProject
 import com.blackbaud.templates.project.PerformanceTestsProject
 import com.blackbaud.templates.project.RestProject
 import org.junit.Rule
@@ -387,9 +388,10 @@ class TemplateGenerationSpec extends AbstractProjectSpecification {
     def "should add consumer pact spec"() {
         given:
         BasicProject basicProject = initBasicProject()
+        PactProject pactProject = new PactProject(basicProject)
 
         when:
-        basicProject.addConsumerPact("foo", "Foo", false)
+        pactProject.addConsumerPact("foo", "Foo", false)
 
         then:
         greenwashOrAssertExpectedContent(basicProject, "add-consumer-pact-spec")
@@ -398,9 +400,10 @@ class TemplateGenerationSpec extends AbstractProjectSpecification {
     def "should add SAS consumer pact spec"() {
         given:
         BasicProject basicProject = initBasicProject()
+        PactProject pactProject = new PactProject(basicProject)
 
         when:
-        basicProject.addConsumerPact("foo", "Foo", true)
+        pactProject.addConsumerPact("foo", "Foo", true)
 
         then:
         greenwashOrAssertExpectedContent(basicProject, "add-consumer-pact-sas-spec")
@@ -409,9 +412,10 @@ class TemplateGenerationSpec extends AbstractProjectSpecification {
     def "should add provider pact spec"() {
         given:
         BasicProject basicProject = initBasicProject("some-service")
+        PactProject pactProject = new PactProject(basicProject)
 
         when:
-        basicProject.addProviderPact()
+        pactProject.addProviderPact()
 
         then:
         greenwashOrAssertExpectedContent(basicProject, "add-provider-pact-spec")
